추가 수정 PostBlockReq의 imageBlock 제거 및 req, res 필드 추가 이름 수정

+++ src/main/java/org/example/chain/domain/post/entity/PostBlock.java
@@ -37,7 +37,7 @@ public class PostBlock {
 
 
     @Column(name = "sort_order", nullable = false)
-    private int sortOrder;
+    private Long sortOrder;
 
 
 }
